Move ThroughputSOSS and WebbKernel from soss_engine.py to engine_utils.py
